chore: remove all depecrated items
