+++ packages/server/index.d.ts
@@ -4,10 +4,6 @@ export type Context<S, C, R> = import("koa").ParameterizedContext<S, C, R>;
 export type Middleware = import("koa").Middleware;
 export type Next = import("koa").Next;
 export type BodyParserPair = import("./src/middleware/body").BodyParserPair;
-export {
-  createBodyParsers,
-  sendFile,
-  compose,
-} from "./src/middleware/index.js";
+export { createBodyParsers, compose } from "./src/middleware/index.js";
 export { closeTestApp, createTestAppAndClient } from "./src/testing.js";
 //# sourceMappingURL=index.d.ts.map
