+++ packages/server/index.js
@@ -20,9 +20,5 @@
  */
 
 export { getApp } from "./src/app.js";
-export {
-  createBodyParsers,
-  sendFile,
-  compose,
-} from "./src/middleware/index.js";
+export { createBodyParsers, compose } from "./src/middleware/index.js";
 export { closeTestApp, createTestAppAndClient } from "./src/testing.js";
