+++ packages/server/src/middleware/index.d.ts
@@ -5,5 +5,4 @@ export { defaultHeaders } from "./headers.js";
 export { healthHandler } from "./health.js";
 export { logMiddleware } from "./log.js";
 export { notFoundHandler } from "./notFound.js";
-export { sendFile } from "./sendFile.js";
 //# sourceMappingURL=index.d.ts.map
