+++ packages/server/src/middleware/index.js
@@ -5,4 +5,3 @@ export { defaultHeaders } from "./headers.js";
 export { healthHandler } from "./health.js";
 export { logMiddleware } from "./log.js";
 export { notFoundHandler } from "./notFound.js";
-export { sendFile } from "./sendFile.js";
