+++ packages/store/index.d.ts
@@ -1,20 +1,16 @@
 export { structure as storeStructure } from "./src/generated/common/structure.js";
 export { queries as storeQueries } from "./src/generated/database/index.js";
 export { setStoreQueries } from "./src/generated.js";
-export { FileCache } from "./src/file-cache.js";
 export { sessionTransportLoadFromContext } from "./src/session-transport.js";
 export { postgresTableSizes } from "./src/insight.js";
-export { sendTransformedImage } from "./src/send-transformed-image.js";
 export type Postgres = import("./types/advanced-types.js").Postgres;
 export type S3Client = import("@aws-sdk/client-s3").S3Client;
-export type MinioClient = import("./types/advanced-types.js").MinioClient;
 export type QueryPart<T> = import("./types/advanced-types.js").QueryPart<T>;
 export type QueryPartArg = import("./types/advanced-types.js").QueryPartArg;
 export type Returning<
   Type,
   Selector extends string[] | "*" | undefined,
 > = import("./types/advanced-types.js").Returning<Type, Selector>;
-export type GetStreamFn = import("./src/send-transformed-image").GetStreamFn;
 export type SessionStoreSettings =
   import("./src/session-store.js").SessionStoreSettings;
 export type SessionTransportSettings =
@@ -24,15 +20,6 @@ export {
   generatedUpdateHelper,
   generatedQueryBuilderHelper,
 } from "./src/generator-helpers.js";
-export {
-  newMinioClient,
-  minio,
-  ensureBucket,
-  removeBucket,
-  listObjects,
-  removeBucketAndObjectsInBucket,
-  copyAllObjects,
-} from "./src/minio.js";
 export { newPostgresConnection, postgres } from "./src/postgres.js";
 export {
   objectStorageGetDevelopmentConfig,
@@ -45,6 +32,8 @@ export {
 export {
   fileCreateOrUpdate,
   fileFormatMetadata,
+  fileVerifyAccessToken,
+  fileSignAccessToken,
   fileSyncDeletedWithObjectStorage,
 } from "./src/file.js";
 export {
@@ -62,15 +51,6 @@ export {
   getMigrationsToBeApplied,
   runMigrations,
 } from "./src/migrations.js";
-export {
-  createOrUpdateFile,
-  fileFormatResponse,
-  copyFile,
-  getFileStream,
-  syncDeletedFiles,
-  fileSignAccessToken,
-  fileVerifyAccessToken,
-} from "./src/files.js";
 export {
   jobFileCleanup,
   jobFileGeneratePlaceholderImage,
