+++ packages/store/index.js
@@ -7,10 +7,6 @@
  */
 
 /**
- * @typedef {import("./types/advanced-types.js").MinioClient} MinioClient
- */
-
-/**
  * @template T
  * @typedef {import("./types/advanced-types.js").QueryPart<T>} QueryPart
  */
@@ -23,10 +19,6 @@
  * @template Type
  * @template {undefined | "*" | string[]} Selector
  * @typedef {import("./types/advanced-types.js").Returning<Type, Selector>} Returning
- */
-
-/**
- * @typedef {import("./src/send-transformed-image").GetStreamFn} GetStreamFn
  */
 
 /**
@@ -49,16 +41,6 @@ export {
   generatedQueryBuilderHelper,
 } from "./src/generator-helpers.js";
 
-export {
-  newMinioClient,
-  minio,
-  ensureBucket,
-  removeBucket,
-  listObjects,
-  removeBucketAndObjectsInBucket,
-  copyAllObjects,
-} from "./src/minio.js";
-
 export { newPostgresConnection, postgres } from "./src/postgres.js";
 
 export {
@@ -73,9 +55,8 @@ export {
 export {
   fileCreateOrUpdate,
   fileFormatMetadata,
-
-  // fileVerifyAccessToken,
-  // fileSignAccessToken,
+  fileVerifyAccessToken,
+  fileSignAccessToken,
   fileSyncDeletedWithObjectStorage,
 } from "./src/file.js";
 
@@ -98,20 +79,9 @@ export {
 } from "./src/migrations.js";
 
 export {
-  createOrUpdateFile,
-  fileFormatResponse,
-  copyFile,
-  getFileStream,
-  syncDeletedFiles,
-  fileSignAccessToken,
-  fileVerifyAccessToken,
-} from "./src/files.js";
-export {
   jobFileCleanup,
   jobFileGeneratePlaceholderImage,
 } from "./src/files-jobs.js";
-
-export { FileCache } from "./src/file-cache.js";
 
 export {
   queueWorkerAddJob,
@@ -144,5 +114,3 @@ export {
 } from "./src/query.js";
 
 export { postgresTableSizes } from "./src/insight.js";
-
-export { sendTransformedImage } from "./src/send-transformed-image.js";
